moved the msg group calls to js file, remove unsed code
