+++ src/ServicePulse.Host/vue/src/components/failedmessages/FailedMessageGroupNoteDelete.vue
@@ -1,12 +1,13 @@
 <script setup>
     const emit = defineEmits(["delete", "cancel"]);
 
-const settings = defineProps({
-    groupid: String
+const settings1 = defineProps({
+        groupid: String,
+        comment: String,
 });
 
 function deleteNote() {
-    emit("delete", settings.groupid);
+    emit("delete", settings1.groupid);
 }
 
 function close() {
@@ -20,11 +21,13 @@ function close() {
       <div class="modal-container">
         <div class="modal-header">
           <div class="modal-title">
-            <h3>Are you sure you want to delete this note?</h3>
+              <h3>
+                  Are you sure you want to delete this note?
+              </h3>
           </div>
         </div>
         <div class="modal-body">
-          <p>Deleted note will not be available.</p>
+            <p>Deleted note will not be available.</p>
         </div>
         <div class="modal-footer">
           <button class="btn btn-primary" @click="deleteNote">Yes</button>
@@ -34,7 +37,6 @@ function close() {
     </div>
   </div>
 </template>
-
 <style>
 .modal-mask {
   position: fixed;
@@ -47,12 +49,10 @@ function close() {
   display: table;
   transition: opacity 0.3s ease;
 }
-
 .modal-wrapper {
   display: table-cell;
   vertical-align: middle;
 }
-
 .modal-container {
   width: 600px;
   margin: 0px auto;
